Exercises/Exercise-5/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.3
sqlalchemy==2.0.23
